+++ frontend/src/pages/Billing/Billing.tsx
@@ -1,38 +1,16 @@
 import React, { useState } from 'react';
 import Stripe from './index';
 import { Container } from '../../components/index';
-import styled from '@emotion/styled';
+import Open from './Open';
 
-interface MyConfig {
-	onClick?: () => void;
-	title?: string;
-	content?: string;
-}
-
-
-const Open = ({ onClick, title, content }: MyConfig) => {
-	const [open, setOpen] = useState(false);
-
-	return(
-		<div>
-		{open ? (
-			<button onClick={() => setOpen(true)}> {title}{content}</button>
-			) : (
-			<button onClick={() => setOpen(false)}> {title}</button>
-			)
-		}
-		</div>
-	)
-}
 const Billing = () => {
-
 	return(
 	    <Container>
 			<div>
-				<Open 
-					title={`first title`}
-					content={`first content`}
-				/>
+				<Open
+					title= {`first title`}
+					content= {`first content`}
+				/>	
 			</div>
     	</Container>
   );
